move class and fix
